refactor: share snapshot rendering across diff and output
